Make start_date deferral for paid periods optional
This adds a defer_payment_start field to membership_payment. If the
field is set to 1 (default), new mandates created for existing
memberships will respect already-paid membership periods and use a
start date after the paid period. This was the previous behaviour.

If the value is set to 0, contract modifications will not respect
paid periods and use the current date for the mandate's start date.